[ISSUE #4903] try use bodyValue
